add tests! inspired by @Ooblioob
This is a fun-sized "project" for experimenting with mocking strategies!

I'm not sure my mocking strategies are ideal; in particular, I'd rather
not use `with mock.patch` ... `mock_open` to deal with testing code that
uses a context manager. And I'd rather not have such blocky
repetitive-looking code in the `setUp` part of `TestLoad`. Hmm.